Stack allocated lvars, GC fixes, compiler changes, oh my!
This is a biggy (too big in fact). It started as a change to allow
arguments to be accessed directly from the stack, and turned into a
monster.

Arguments and some lvars can now be accessed directly from the stack,
making them cheaper to create and use. This turned out to expose
a large number of bugs in the VM related to stack access, as well
as some in the GC.

The big GC change here is that the mark/sweep GC is actually run now,
as opposed to before when it would just allocated more and more memory
(the source of memory issues I suspect).