The context optimization change your parents always talked about but never implemented.
Is now implemented!

The biggy is here is the change to have context's are allocated. They're now allocated in a stack fashion for easy reuse, and automatically converted to full objects each run of the garbage collector as needed. Only contexts which have been referenced are saved at collection.

Also in here is a little cleanup to object initialization. Object memory was clearing the body of an object 2 times too many, so I restructured the code so it's delayed until it really needs to be done. Also, added memutil.h, which contains some simple but fast memcpy and memset substitutes that only function on memory that is word aligned (all objects are, so thats easy.)