Updates to bin/completeness and bin/mspec.
Added -x excludes to completeness to enable it to be
used with rubinius and spec/reports/critical-excludes.txt.

Clarified args to -x, -e for bin/mspec. Args must be a string
or file. If a file, each line is equivalent to a -x instance.

Removed legacy scripts from bin. Removed old reporting from bin.
Completeness and rspec reporting was broken and self-hosting
reporting is obsolete. We are self-hosting.