Finished context and cpu simplification
CPU only cache's 5 things from the context now, and doesn't use a memcpy
to transfer data between the active_context and the cpu.

Fixed a bug in the compiler where the line numbering for the callsite
of a block was off.

Block calling is faster now, as the block_call primitive handles the args
directly now.