Ticket #65: Compile {...} to call Hash.[].
This means we're now calling key.hash rather than relying on
object_hash_int() to get the key's hash value.

I changed the order in which the hash elements are taken out of
the argument array so that the order of the hash elements equals
the order in which the elements are specified.

The make_hash opcode is now unused.