Make #is_a? exactly the same as #kind_of?. See explanation below.
The #is_a? method is an alias of #kind_of?. We have a compiler
plugin that handles calls to #kind_of? with a special operation.
Since the programmer should not have to distinguish semantically
between #is_a? and #kind_of?, she should also not have to distinguish
the execution characteristics either. I added #is_a? to the compiler
plugin as well. This is obviously not DRY, but I think that's OK.