Fix generator breakages to ensure_return and regex literals
Also added back cm.primitive = @primitive line, but left this
commented out, as it breaks the ability to execute code compiled
under compiler2. Appears to be a mismatch between use of @primitive
in compiler2 (@primitive is set to names of primitives) and
CompiledMethod (which expects a primitive index). Not sure which
usage is correct.