Fix for compiler2 optional args with block default
Workaround an apparent bug in ParseTree, whereby a def that contains
an optional argument whose default value is a block gets the wrong
argument name for the optional arg. For example,
  def foo(output = 'a', prc = lambda{|n| output * n})
gets converted to the args sexp
  [:args, [], [:output, :n], ...
when it should be
  [:args, [], [:output, :prc], ...

With this fix, spec/language/def_spec.rb now compiles and successfully
passes under compiler2.