Tighten spec in core/hash/equal_value_spec.rb
"Hash does not compare keys with different hash codes via eql?" was
failing because it detected that eql? was called on a key.  However,
eql? was not being used to compare keys with different hash codes
from the two hashes.  Instead, eql? is used to compare a key to itself
during hash element reference, in order to distinguish between two keys
with equal hash codes that aren't eql?.

The tightened spec only fails if the keys are compared eql? to each other.