Fix SHA1 on big endian platforms
This code originally used LITTLE_ENDIAN as a macro it #ifdef
checked to decide how to rotate some code. The problem is that
LITTLE_ENDIAN is a unix system macro, that is used to make
'#if BYTE_ORDER == LITTLE_ENDIAN' work.

Just using a better macro name fixes the bug.