Added control over backtrace colorisation
Default colorized backtrace shows the last context
in red. This is fine when backtrace is due to an
exception, but less appropriate when the backtrace
is used for debugging to show the current location.
Added ability to set the colors used when showing
the backtrace, as well as disabling colorisation
on request.