A new algorithm for keeping track of the statuses of children processes
The old algorithm involved calling waitpid(-1, &status, WNOHANG) both
from a SIGCHLD signal handler and from the Process.waitpid primitive.
Child statuses were being saved in a hash when the children dies so
that they could be reported for later calls to the primitive.
Essentially, it eagerly reaped all children whenever the died, and
saved their exist statuses if no one was waiting for that child yet.

Aside from making it difficult to implement the waitpid(-N) and
waitpid(0) cases (since the process group id is not available after the
child has died), the fundamental problem with this algorithm is that
it requires atomically checking the hash table for a previously reaped
child matching the pid supplied and conditionally adding a thread_info
to the thread_infos list, with no possibility for any intervening
SIGCHLD signal handling.

The new algorithm is the opposite extreme in some senses.  It never
calls waitpid() unless a Process.waitpid has been called, even in
the SIGCHLD signal handler.  Instead, for each call to Process.waitpid,
we save all the arguments and then manually call the SIGCHLD handler.
In the handler, we only call waitpid() with the arguments passed to
Process.waitpid (but always with WNOHANG to avoid blocking).  Therefore,
whenever waitpid finds a dead child, we can always send the result to
the channel immediately, so we don't ever have to save the result
anywhere.

The new algorithm avoids the race condition by not maintaining any
hash table of previously reaped children and avoids the complex
storage and matching of process groups by letting the system's waitpid()
handle it all.

Data structure and other changes:

Added a 'type' field to the thread_info struct so that cpu_find_waiters()
can ignore the other thread_infos in the list.

Added an 'options' field to the thread_info struct so that we can
record the 'flags' argument to Process.waitpid, and avoid waiting for
the child to terminate when WNOHANG is used.

Removed the recent_children hash from state, since it's not needed
anymore.