Fix for leaking of signal handler threads.
Previously, a signal handler registered with Signal.trap was not
removed when a new handler was registered.  This left the thread and
the thread_info struct around indefinitely.

Now, before registering the new handler we send nil the channel
waiting for the old handler, so that it knows to die.  And, we unregister
the signal with libevent, and free the thread_info struct.

Note: no spec exists for this yet, but this commit prevents
"Warning: encountered invalid class (not big enough)" errors in other
specs that use Signal.trap.