Add dependencies necessary to build with JRuby running rake.
It's not quite perfect yet, but a build under JRuby now runs as expected
and bin/ci appears to have the same or fewer failures with a JRuby-built
version as with MRI-built pre or post dependency changes.

The remaining problem is that the build doesn't appear to successfully
compile extensions. I will continue looking into that, but it appears
that it may be an issue with load orders in perhaps platform or
bootstrap.