Fix the detection of numbers too big for Fixnum.
When comparing the result of rbs_to_int() (which is typed as long)
to another number to see if that number would fit into a Fixnum,
you must cast the result to (int), since int may be smaller than long.

Also, you must avoid comparing unsigned types with the signed result
from rbs_to_int() since the unsigned value may be coerced to a signed
type, making the comparison true even when the original values differ
in sign.

Also fix a related bug in ffi_from_ll():

Casting the long long to a int makes it impossible to detect differences
from the int in the upper bytes.