Fix inspect spec to guarantee the target thread is actually sleeping.
A reminder for folks adding Thread specs: You *CAN NOT* know that a target
thread is sleeping unless you are polling for status == 'sleep'. No amount
of channel, lock, or state variable tricks will get around that. Please
use polling if you want to guarantee a target thread is asleep.