Modified yield_debugger opcode to send thread
For debugging code with more than one thread, it is necessary
to know the thread that hit the breakpoint, in addition
to the execution context. As the context can be retrieved
from the thread (via the thread's task), we now pass the
thread on the debug channel instead of the method context.