Remove the goofy Hash#delete spec and replace with two others.
The old version of this example depended on individual hash buckets having a
specific ordering, which overreaches a bit. The new version, while a little
cumbersome, should work correctly regardless of hash implementation or hash
and bucket ordering.

I also moved out a few lines that were unrelated to this example into a
separate one.