Modified yield_debugger to use default debug/control channels
If a thread/task has not had debug and/or control channels set
via a call to set_debugging, when a yield_debugger opcode is
encountered, a VM-wide debug channel will be used if one
exists, and a control channel will be created on the thread
so that the thread can be resumed.

This removes the need to call set_debugging on every thread
that might encounter a breakpoint, while still maintaining
that capability should specific channels need to be set.