Modify yield_debugger to not set debug_channel
Modified yield_debugger to not set the debug_channel to
the VM-wide debug channel when the thread debug_channel
is nil. This prevents threads that have not previously
had set_debugging called on them from subsequently being
bound to the VM-wide debug channel that was set at the
time they hit a breakpoint.

Also modified logic so that a control channel is only
created if a valid debug channel has been located.