Fix libev bug for signal handling after forks.
We must not reinit an active sigev watcher.
The old code was reinitializing the sigev watcher inside a
generic loop_init() function, which was used both for
default loops and non-default loops.  That meant that the
process of creating a non-default loop was wiping out the
active count of the global sigev watcher.

This commit moves the sigev initialization into the default-
loop-only initialization.