Fix libev bug - losing signals while resetting non-default loops.
After a fork, the pipe used for signal detection needs to
be closed and reopened.  This creates an unavoidable hole
during which signals may not be received immediately after a
fork.  However, this only needs to be done once.

The process of resetting the non-default loops after a fork
does not need to close and reopen the signal detection pipe,
thus creating more opportunities for signal loss.