Fix libev bug so fork-fixups work even with no active events.
The event loop was checking for active event watchers before
checking for the need for a fork-fixup, so it was bailing
out early, leaving the child dropping all signals until
after a watcher was installed and the event loop rerun.

This fix just checks for the need for a fork-fixup before
checking for active watchers.