Fix a signal handler race in libev.
The race:
  Assume there are two signal watchers started, and gotsig is 0.
  The first signal is received, triggering sighandler().
  gotsig becomes 1.
  The signal handler returns, so signals are unmasked.
  Before sigcb() clears gotsig, a new signal is received.
  This signal will be lost because of the "if (!gotsig)"
  condition.

The fix:
  Get rid of the gotsig variable altogether.

Note:
  I don't really know what this variable was for, so I don't
know if it was meant to fix some other problem.  My best guess
is that it was an attempt to compensate for buggy OSes that
called the sighandler multiple times for one signal.  If there
are such systems, this will no longer hide their behavior.
(Which it wouldn't have completely done before, either.)