Add support for setting a context iseq directly
A context can now be set to use a specific instruction
sequence directly, instead of using the iseq from the
compiled method it is currently executing. This capability
allows the debugger to swap in and out yield_debugger
instructions for the purposes of stepping and re-applying
persistent breakpoints, actions which should only modify
the instruction sequence for a single thread/context.