Added yield_debugger on context change
Added capability to set a flag in the VM that causes a
yield to the debugger to occur immediately following a
change in the active context. This provides a foundation
for step in logic for the debugger, which need only set
a flag on a task and have a breakpoint triggered at
whatever receiver is activated following a send.