Simplify Dir#open spec dependencies.
The "takes a block which yields the Dir instance and closes it after"
expectation was failing because it depended on File.for_fd working
properly with closed file descriptors which it does not.

This revision removes the dependency on File.for_fd as well as
IO.sysopen (which is not yet implemented in jruby) such that
the spec now passes on ruby, rbx and jruby.

Signed-off-by: Eero Saynatkari <rubinius@projects.kittensoft.org>