Disable Large File Support (LFS) for now
Enabling it breaks Dir#read, because we expect a LFS DirEntry, but it looks
like we get a non-LFS version. This is really strange, because everything
should be compiled with LFS enabled.

If we choose to enable LFS for Linux, I think we should also do it for OS X.
This basically comes down to using _DARWIN_USE_64_BIT_INODE.