Fix the build on Linux. See the discussion with be9 on IRC
This is not a real bugfix, but hides the problem for now. The problem is that
enabling DISABLE_CHECKS en disabling EXTRA_PROTECTION exposes bugs in Linux.

This is not related to the Dir / Stat problems, but seems like a completely
separate problem. It was exposed because in a56104f1 DISABLE_CHECKS was set
to 1. The last time this value was changed was on October 5th, so somewhere
between than and now this problem was introduced.

On 64 bit, EXTRA_PROTECTION is also needed to prevent segfaults. Because
of a small typo in state.h (ifdef EXTRA_PROTECTION instead of
if EXTRA_PROTECTION), this was not exposed. When DISABLE_CHECKS was set
to 0, the EXTRA_PROTECTION code was also run.

Hopefully someone can isolate the cause of this bug. We've tested with
different distributions and difference GCC versions on Linux, but the
bug is consistent over all distro's.