Fix FFI so (unsigned) long long values are not truncated on 32 bit platforms
The conversion inside FFI was broken, because the return value from nf_converter
was defined as long. This is changed to uint64_t, which is the largest type
we currently use.

The functions that use the nf_converter type have been verified to properly
cast smaller values back so they still work correctly.

This change also allowed for Process::Rlimit to be converted to the new Struct
generator style. In the old version, it used incorrect offsets too, so it's a
miracle that the process specs even worked ok.

Theoretically these fixes also fix possible problems with other 64 bit types
on 32 bit platforms, such as dev_t and ino_t, which would have been truncated.