Prevent GC from resetting customised iseq on context
The debugger relies on the ability to customise the iseq
in use on a context so that a context being debugged can
use a different compiled iseq than other contexts running
the same method. This allows the debugger to insert and
remove yield_debugger instructions on just the debugged
context, without those changes being seen by e.g.
different threads.

This patch fixes an intermittent problem whereby the GC
would reset a context's compiled iseq to the associated
CompiledMethod compiled field, thereby putting the context
into an unexpected state and messing up the debugger.