Reduce over-use of Range in some core classes. See below.
On a full CI run, Range objects accounted for ~7.7% of all
objects created by calls to the allocate primitive. Additionally,
the following is a list of the caller and number of calls to
Range methods:

124,008   Regexp#pre_match_from
 35,271   Dir.glob_helper
 24,516   Regexp#matched_area
 20,520   Regexp#post_match
 17,688   Regexp#pre_match
 15,378   Regexp#get_capture
 12,444   String#gsub
  5,115   Sprintf#parse

In all these cases, Range instances were being used when a
simple start, length was either readily available or easily
calculated.