Fix coercions for File#chmod
In order to fix this, the FFI conversions are also changed.
It is now possible that even a Bignum ends up being a short in
C land, this is done by simply casting it. This was already happening
for int variables larger then 2**16 - 1 and this is also the same
behavior that MRI exposes.