Fix for profiler non-deterministic failures (LH #365)
Shotgun profiler was failing randomly, due to the SIGPROF
signal being received while a method context was in the
process of being freed. We now clear the active context
before retiring it, so that if the profiler samples the
current context while we are freeing it, it gets nil.
This prevents the sampler from grabbing a context that
will be freed, leading to corruption.