InstructionSequence#decode now strips trailing noops
ByteArrays (and hence ISeqs) are allocated in chunks that
are 4- or 8-bytes in size. This means an iseq can have
trailing noops at the end to pad the length to a 4- or 8-
byte boundary; these noops were appearing in the decoded
This in turn caused functions such as line_from_ip to fail
if passed one of these trailing noops.

ISeq#decode now strips these trailing noops.

Also changed iseq_new to take a size in bytes, rather than
fields, which is confusing (since iseq_new is called by
cpu_marshal and was dealing in fields, but ISeq.new
inherits new from ByteArray, which deals in bytes).