Add Module.allocate primitive.
There is no good reason to (and many good reasons not to)
duplicate in Ruby what the VM already provides. In this case,
Module.allocate was implemented in kernel/core when the logic
is already provided in the VM. This severely complicated
converting constants to use LookupTable rather than Hash.