Reluctantly introduce load order dependencies to bootstrap. See below.
LookupTable needs to be fully set up in bootstrap so that methods such
as #key? and #has_key? are available in class bodies in kernel/core.
For example, gem_prelude.rb has this line at script level

  if defined?(Gem::Enable) && Gem::Enable then

which ultimately calls #has_key? on a constants_table, which is an
instance of LookupTable, but #has_key? is an alias of #key?

The alternative to introducing these load order dependencies in
bootstrap is to require that only one of #key?, #has_key?, #include?
or #member? be used when coding methods which *may* be called by other
class body or script body code in core. That would seem to be way too
big a gotcha compared to load order dependency in bootstrap.