Add notion of eval context to Debugger, plus up/down cmds
The evaluation context in the Debugger is no longer bound
to the debug context, but can be moved up or down the call
stack using the new commands up and down.

Commands such as eval, list, vars, ivars, etc use the eval
context rather than the debug context, so you can use up
to move to the caller, and then l to view the source from
the call site, vars to inspect the value of locals at the
call site etc.