Replace bm_string and scale all benches to approx same magnitude.
This scaling has at least two benefits: 1. the amount of work done
in each bench is approximately the same, so fluctuations should be
approximately the same also; 2. it provides an explicit comparison
for the relative costs of different operations on MRI and, to the
extent that these differ on RBX, gives another view of RBX execution
profiles.