Modify Kernel#debugger to remove send from call site
After Kernel#debugger is called, it modifies the call
site to replace the send with a noop. This means the
debugger method is only called once from any location,
so that if it appears in a loop, say, it only activates
the debugger once, and quitting the debugger really
quits (unless a new debugger call is encountered
subsequently).