Fixes and specs for InstructionSequence::Encoder#replace_instruction
Generally replace_instruction doesn't replace an instruction which is
wider than the original one. A special case is when the original
instruction is followed by enough noop instructions. It can then be
safely replaced. The code before this patch can't handle this case
(though it was intended to handle it).

This commit includes specs for this case (and other cases) for
InstructionSequence::Encoder#replace_instruction and fixes the problem
mentioned above.