Implemented multi-process option for mspec/bin/mspec script.
Usage: bin/mspec ci -j

Note: This facility is only intended to run the set of CI files
with a specific formatter. The behavior is *undefined* if you
pass anything besides -j. So, please don't complain. :)

The facility is implemented in the mspec script because you can
use MRI to run this script and launch the subprocesses with the
intended target implementation. This facility does not use
Kernel.fork on JRuby or (MRI) Windows.

The primary purpose of the facility is to run e.g. the core specs
independently from the stdlib specs but still have the appearance
of a single process.