Add support for conditional breakpoints
Conditional breakpoints can now be used in the debugger.
The special variable #hits can be used in a condition that
depends on the number of times a breakpoint has been hit, e.g.

  b foo/bar.rb:10 if #hits > 10

The condition is evaluated in the context of the breakpoint,
so local vars, self, ivars etc from the breakpoint context
can be used in the condition expression.