Run spec:update before and after git:update.
According to Evan, both before and after are necessary. If you
run into a problem doing rake git:update, note whether there is
a change to the submodule commit for either mspec or spec/frozen
in the changes you are trying to pull. If there is, let us know
so we can further tweak this process.