Add AccessVarMethod#name and DelegatedMethod#name
UnboundMethod#inspect and Method#inspect assume a #name on the @compiled_method
instance variable, which could be a CompiledMethod, an AccessVarMethod or a
DelegatedMethod. Since irb calls #inspect, evaluating things like
"obj.method(:whatever)" in irb will throw an exception without this change if
@compiled_method one of the latter two types.

To make AccessVarMethods appear as "Class#var" instead of "Class#@var", I added
an extra name parameter offloading that work to the caller.

DelegatedMethods appear as Class#<delegated to XXX>.