Fixed a bug of FFI loading function from specific library.
library.respond_to? :each should not be used here because
String.respond_to?(:each) will be true. Though currently
String.each will iterate line by line which will not cause
problem here. But this will be a potential bug if the API
of String iteration is changed according to Ruby 1.9. So
a guard of String is added here.

Further more, the final "return nil" is needed. Or else
some garbage value (e.g. the library name) will be returned.