Initial implementation of ARGF (with some updates).
The implementation is fine generally. Converted from using
class methods to a simple class with a single instance, the
same way ENV is done. Added $< manually since that part of
the patch would not apply. Also changed the

  private def xxx

to normal 'private :xxx' following the method definition
so the code is MRI compatible.

Signed-off-by: Brian Ford <bford@engineyard.com>