Fixed sydney parsing def(a=1, b=2, c=lambda {|n| n}).
Basic problem was assuming contents of locals array had the names
of the optional args in the same order they appear lexically. This
is untrue in the case above because the lambda is processed before
the c=. Thanks to Evan for pointing out the problem and explaining
points about the node structure.