Possible fix for the unspecified zone problem
Our current argument handling code creates a context, then allows a
specialized class to validate the passed in args and setup the context.
The problem is that if the class says the args are bad, we throw an
internal argument error right away. This leaves the sender field of the
context uninitialized.

This is fixed by always initializing sender to nil in init_context.
Also, the other fields of MethodContext have comments above them,
indicating where they are initialized.