Revert "Use Comparable and spaceship to prevent infinite Numeric looping"
This reverts commit 19fba41df92281b5821a390c619f412ab4076f02.

The way that the numeric tower is implemented in Rubinius, each basic
class (e.g. Fixnum, Bignum) implements specific versions of methods
like #== as primitives. If the primitive code fails (typically
because it is unable to coerce to a basic type -- see docs on primitives
in general), the method on the basic type delegates 'up' with a call to
super.