reorganize uncache_stack calls
based on 3 possibly flawed rules. save_stack_pointer code should
only be emitted [1] before ret or [2] before function call that's
passed the ctx [3] and when the stack is cached.

does the stack uncache after 'if(last_imm) {' mean that we don't
want to honor any caller changes to the stack pointer in ctx?