Implement exception handling using a dynamic unwind stack
The exception table is no longer used. Instead the setup_unwind
and pop_unwind instructions are used to register and unregister
information about what to do when an exception occurs.

The unwind information includes a target ip to jump to and a
target sp to set the stack to when delivering the exception.

This properly allows us to implement begin;rescue;end as an
expression.