A dup'd context's home_ should be referenced.
Say dup = ctx->dup(); then ctx->home (same as dup->home()) should be
referenced so that it won't be deallocated. Otherwise, the closure
may be lost. One example is kernel/common/continuation.rb: when a
continuation is restored, the local variable cont might point to an
invalid address and cause a NoMethodError.