Introduces the dynamic interpreter
The dynamic interptere uses the same assembly generation infrastructure
as the JIT, but instead it writes an interpreter loop into memory to use
instead of the static interpreter loop.

This dynamic interpreter is currently a little faster than the static
interpreter, with a bunch more places for optimization.

It also opens the door for being able to generate interpreters that
perform profiling that the JIT later uses.