Merged BlockEnvironment methods in common/context.rb into block_environment.rb
Note that context.rb was shadowing the implementation in
block_environment.rb.

In particular, make_independent had a different implementation, which I
merged to the best of my abilities to avoid spec failures.

Additionally, the use of @bonus for @metadata_container is deprecated
but was being revivified by the shadowing BlockEnvironment implementation
in context.rb.