Converted all parser specs to new form.
This is the first step to reworking the compiler specs. Once the
bytecode expectations are added, the specs will be moved to the
spec/compiler directory.

The specs are organized most directly by AST node. This can follow
Ruby source very closely (e.g. 'yield' and :yield node), or it can
be superficially different (e.g. '1 while not a' results in an :until
node). The underlying AST similarity is more important for clarity
in the specs. Another issue is how to organize nodes like :retry and
:redo. All the iterator (block) specific nodes are under iter_spec.rb.
If this becomes unweildy, they can be split out.