Fixed TestIO so it doesn't leave dead files in /tmp.
We can safely unlink the temporary file after it was opened since it
won't actually go away until the file descriptor is closed as well.
I also got rid of the superfluous strdup() call -- mkstemp() needs a
writable buffer, but it doesn't matter whether it's not the stack or
on the heap.