Migrated the remaining old compiler specs.
Added the specs for compiler classes to spec/compiler/stages. The node
specs are much more extensive and useful. Instead of putting them in
spec/nodes, they remain in spec/compiler. The specs for the compiler
classes are in spec/compiler/stages instead of mixing them in with the
node specs.

If this is really confusing or unwieldy, we can revisit. However, this
maintains a good balance between the primacy of the node specs while
still keeping everything organized simply under spec/compiler.