Fixed issues with Message not being initialized.
A couple specialized constructors were added that minimize the number
of initializations that are then assigned to. This ensures that the
Message instance is fully initialized before being used and avoids
an extra call to Message::setup. Effective C++ suggests the compiler
may be able to emit better code in this case. Overall, the code for
the send_* instructions is significantly cleaner.