Mongrel installs/builds on OS X. Removed mongrel ext.
There may be issues with rbconfig.rb and mkmf.rb on
various platforms. These issues will be addressed as
they arise. The approach is to check the Rubinius-
generated rbconfig against the MRI one for that
platform and fix the Rubinius one as needed. There
is probably not a need to modify mkmf.rb.