Remove filtering methods in profiler output.
The reason filtering was added is because there are a *lot* of Rubinius-
specific methods (outside of the core library classes like Array, etc.)
in the output and that can be noisy for a user.

However, filtering skews the profile results, so I removed it. Also, if
the implementation classes/methods are ugly, they should be refactored/
cleaned up rather than hidden.