Rework capi Handles entirely
Handles are now full C++ objects and are integrated with the GC. Handles
contain a reference count to indicate if there are any strong references
to them, ie, should the Handle keep the object alive.

If the Handle is weak, it will not attempt to keep the object alive, but
it will make sure that it stays up to date with the object as it
continues to live.

This allows us to have constant memory location Handles for objects that
move around, without having handle temporal locality problems like the
previous implementation.