Changed profiler cumulative seconds method.
The profiler now uses two timers to gather stats. One timer, the StackTimer,
maintains a running wall-clock time during the invocation of a method, even
if the method is re-entered. The other timer tracks the time spent in each
invocation of a method and each callee. These timers may be running
concurrently if a method is re-entered. The final self seconds is the total
time spent in all invocations of a method less the time spent in all its
callees.