Moved implementation classes under Rubinius namespace.
The following classes have been namespaced under Rubinius:

ARGFClass                   InstructionSet
AccessVariable              Location
BlockEnvironment            LookupTable
ByteArray                   MachineMethod
Channel                     MethodTable
CompactLookupTable          NativeMethod
CompiledFile                Scheduler
CompiledMethod              Selector
DelegatedMethod             SendSite
EnvironmentVariables        Sprintf
Executable                  StaticScope
GlobalVariables             ThrownValue
Globals                     Tuple
IncludedModule              VariableScope
InstructionSequence

Evan specifically requested to not namespace MetaClass. Backtrace
is also not namespaced. Compiler was not namespaced because it will
be restructured in the near future.

The following method was used:

module Rubinius
  class Tuple
  end
end

rather than

class Rubinius::Tuple
end

Primarily, this was chosen because it allows implementation classes
to refer to other implementation classes without qualifying the name.
This results in less noise in the implementation classes and is less
likely to cause exceptions if someone forgets to qualify a name. In
the regular classes like Kernel, references to implemetation classes
do need to be qualified (e.g. Rubinius::VariableScope in Kernel#eval).
A consequence of this method is that git blame history info is masked
for a number of classes. (If this info is needed, check out a branch
and git reset --hard to this commit - 1.)