SymbolTable::lookup now works with char pointers instead of std::strings.
Before, the only caller of SymbolTable::lookup(std::string s) was
SymbolTable::lookup(char* s), which constructed a std::string to pass
to the former method. Making the method work on char*'s directly allows
us to drop the construction of those temporary std::strings.

This has a small but measurable performance impact for me.