Rework how blocks get their arguments
Rather than always allocating a Tuple and having the block pull it's
arguments out of that, the C++ Arguments object is now passed into the
interpreter. The cast_for_* instructions then read args and put results
on the stack. For passing and receiving a single block arg, the path is
much more straightforward, and much faster.