Remove VariableScope::exitted
Rather than setting this flag, before performing a long return, the
scope is validated by searching up the CallFrame chain to see if the
scope is still running.

This is also safier than then flag, because it doesn't accidentally
allow a long return that is trying to occur between 2 threads.