Improvements to current Hash algorthm.
These experiments are located at doc/experiments/hash/code/chained*.rb.
The code in chained_rev2.rb uses the cleanest code for the Iterator.
Using this code would enable using iteration internally for several
Hash methods thereby removing a layer of blocks. The improvement can
be seen in the spreadsheet on the bm_each benchmark:

  http://spreadsheets.google.com/ccc?key=rdiNYWh0xTGShVHikWVmnUA

However, the Iterator code quality has a tradeoff in performance at
the moment because of the extra read and write to an ivar. This can
be improved with improvements to CompactLookupTable.