Temporarily commiting benchmark graph.
The baseline graphs still need to be added. This shows basically
how the graphing works.

A benchmark dir (e.g. benchmark/app) has several benchmark files.
Each of these files are run ITERATIONS times (defaults to 5). This
produces a min, max, mean, median time for that benchmark. (The
ITERATIONS constant could be changed to SAMPLES or TRIALS.) Each
benchmark may use a set of inputs (typically, the number of times
to loop doing some work), so each benchmark file generally corresponds
to a set of numbers (e.g. some-work-10, some-work-200, some-work-5000).

The benchmarks in a directory are run on a particular VM with certain
settings. This creates a grouping of benchmark results (e.g. results
for MRI, rbx with no jit, rbx with jit, rbx with some X).

The results are graphed over time with each "group" generating a
separate graph. The field (e.g. max, min, etc) is used to extract a
single value for each benchmark + input and that value is graphed
versus an X axis of dates.

The ability to add a baseline to the graphs and another set of graphs
that show a delta to the baseline needs to be added.

Generating these requires 4 basic steps:

1. Run the set of benchmarks with the baseline configuration.
2. Run the set with each of the configurations of interest over
   a series of times (e.g. hours, days, weeks)
3. Run rake -q bench:results FIELD=median
4. View the benchmark/results/web/index.hmts