Added Rubinius::LoadedFeatures for $LOADED_FEATURES.
I contemplated writing a general purpose ArrayMap that does
essentially what LoadedFeatures does for a few methods. But
there are issues with that, primarily what happens if the
Array is extended with mutator methods. The methods called
on $LOADED_FEATURES based on a google code search are few.
So just those methods were implemented. Other methods may
need to be added in the future.