Revisit @total use in Array, see below.
The Array methods can be divided into those that know about the
underlying implementation details and those that do not. It does
not make sense to use #size in place of @total in a method that
also uses @tuple or @start. So, these guidelines are set:

1. If a method deals with implementation details of an Array,
   consistently use the implementation elements (i.e. use only
   @total, @start, etc. not #size, #start. For consistency,
   use the same accessors on the other array (i.e. use only
   other.total, other.start, etc.).
2. For all other Arrays, use the Ruby interface. These include
   objects passed as arguments. In other words, any Array that
   is not being copied from, etc, should be accessed via the
   Ruby method interface as if it is opaque.

Also, in general, code should NOT be written as follows:

  obj.some_method(arg1,arg2,arg3+arg4)

Omit unnecessary parentheses and always put space after a comma
and between the operator and its operands in method bodies. Do
NOT put spaces around the = assignment operator in method
definitions. For example

  def some_method(a=1, b=nil)

NOT

  def some_method(a = 1, b = nil)

NEVER omit parentheses from method definitions that have arguments.