Reworked capi cached data flushing/updating.
The intersection of the set of programs that use RSTRING_PTR(str)
as a read-only buffer and the set that does not is non-empty. So,
to keep the confusion in C extension code to a minimum, Rubinius
supports using the pointer returned by both RSTRING(str)->ptr and
RSTRING_PTR(str) as a read-write buffer (i.e. changes to the buffer
are flushed to the Ruby string, and changes to the Ruby string are
updated in the buffer).

Rubinius defines an alternate API for working sanely with Ruby
strings from C extensions. The functions are:

  rb_str_ptr
      provides a persistent char [] buffer whose contents must be
      explicitly flushed or updated by calling rb_str_flush or
      rb_str_update.
  rb_str_ptr_readonly
      provides a persistent char [] buffer but the contents are
      neither flushed nor updated as for RSTRING.
  rb_str_len
      returns the length of the string without causing caching.
  rb_str_flush
      flushes the cached data to the Ruby string *only* if the
      string is cached.
  rb_str_update
      updates the cached data from the Ruby string contents *only*
      if the string is cached.

This API attempts to accomplish three things:

  1. the intent of the code is clear (i.e. calling rb_str_ptr or
     rb_str_ptr_readonly depending on whether the code expects
     to update the Ruby string after writing to the cache).
  2. performance; the flushing and updating happens when needed
     rather than on every transition between C code and managed
     code.
  3. isolation of the C code from the underlying implementation.

There is a potentially significant cost to using RSTRING(s) or
RSTRING_PTR(s). The cost depends on the number of handles to strings
with cached data and the number of transition calls (i.e. calls
like rb_funcall that must transition from C code to managed code).

The RSTRING macros are provided mainly to simplify running existing
C extension (including gems that have native extension components).
The compat.h file has been update with the above functions and all
C extension authors are encouraged to use this API.